fix: move createCommit (v3) method to stackService
