+++ apps/desktop/src/lib/commits/commitService.svelte.ts
@@ -1,19 +1,7 @@
 import { Commit } from './commit';
 import { ReduxTag } from '$lib/state/tags';
 import { plainToInstance } from 'class-transformer';
-import type { HunkHeader } from '$lib/hunks/hunk';
 import type { ClientState } from '$lib/state/clientState.svelte';
-
-type CreateCommitRequest = {
-	stackId: string;
-	message: string;
-	parentId: string;
-	worktreeChanges: {
-		previousPathBytes?: number[];
-		pathBytes: number[];
-		hunkHeaders: HunkHeader[];
-	}[];
-};
 
 export class CommitService {
 	private api: ReturnType<typeof injectEndpoints>;
@@ -24,12 +12,6 @@ export class CommitService {
 
 	find(projectId: string, commitOid: string) {
 		const result = $derived(this.api.endpoints.find.useQuery({ projectId, commitOid }));
-		return result;
-	}
-
-	// eslint-disable-next-line @typescript-eslint/promise-function-async
-	createCommit(projectId: string, request: CreateCommitRequest) {
-		const result = $derived(this.api.endpoints.createCommit.useMutation({ projectId, ...request }));
 		return result;
 	}
 }
@@ -44,13 +26,6 @@ function injectEndpoints(api: ClientState['backendApi']) {
 				}),
 				transformResponse: (response: unknown) => plainToInstance(Commit, response),
 				providesTags: [ReduxTag.Commit]
-			}),
-			createCommit: build.mutation<Commit, { projectId: string } & CreateCommitRequest>({
-				query: ({ projectId, ...commitData }) => ({
-					command: 'create_commit_from_worktree_changes',
-					params: { projectId, ...commitData }
-				}),
-				invalidatesTags: [ReduxTag.StackBranches, ReduxTag.Commit]
 			})
 		})
 	});
